src/holopy/quantum/error_mitigation.py
"""
Quantum error mitigation with holographic corrections.
"""
from typing import List, Tuple, Dict, Optional
import numpy as np
from scipy.optimize import minimize
import logging
from dataclasses import dataclass
from .noise import HolographicNoise
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT
)

logger = logging.getLogger(__name__)

@dataclass
class MitigationResult:
    """Results from error mitigation."""
    mitigated_state: np.ndarray
    fidelity_improvement: float
    error_reduction: float
    confidence: float
    processing_cost: float

class HolographicMitigation:
    """Implements quantum error mitigation with holographic constraints."""
    
    def __init__(
        self,
        n_qubits: int,
        noise_model: Optional[HolographicNoise] = None,
        learning_rate: float = 0.01
    ):
        """
        Initialize error mitigation system.
        
        Args:
            n_qubits: Number of qubits
            noise_model: Optional noise model
            learning_rate: Learning rate for optimization
        """
        self.n_qubits = n_qubits
        self.noise_model = noise_model
        self.learning_rate = learning_rate
        
        # Initialize mitigation parameters
        self._initialize_mitigation()
        
        logger.info(f"Initialized HolographicMitigation for {n_qubits} qubits")
    
    def _initialize_mitigation(self) -> None:
        """Initialize error mitigation parameters."""
        try:
            # Initialize extrapolation parameters
            self.scale_factors = np.linspace(1.0, 2.0, 5)
            
            # Initialize error maps
            self._initialize_error_maps()
            
            logger.debug("Initialized mitigation parameters")
            
        except Exception as e:
            logger.error(f"Mitigation initialization failed: {str(e)}")
            raise
    
    def mitigate_errors(
        self,
        noisy_state: np.ndarray,
        target_state: Optional[np.ndarray] = None
    ) -> MitigationResult:
        """
        Apply error mitigation to quantum state.
        
        Args:
            noisy_state: State with errors
            target_state: Optional target state for comparison
            
        Returns:
            MitigationResult containing mitigated state and metrics
        """
        try:
            # Apply zero-noise extrapolation
            extrapolated_state = self._zero_noise_extrapolation(noisy_state)
            
            # Apply probabilistic error cancellation
            cancelled_state = self._probabilistic_error_cancellation(
                extrapolated_state
            )
            
            # Apply quasi-probability decomposition
            mitigated_state = self._quasi_probability_decomposition(
                cancelled_state
            )
            
            # Calculate metrics
            metrics = self._calculate_mitigation_metrics(
                noisy_state,
                mitigated_state,
                target_state
            )
            
            logger.debug(
                f"Applied error mitigation with "
                f"fidelity improvement: {metrics.fidelity_improvement:.4f}"
            )
            
            return metrics
            
        except Exception as e:
            logger.error(f"Error mitigation failed: {str(e)}")
            raise
    
    def _zero_noise_extrapolation(
        self,
        state: np.ndarray
    ) -> np.ndarray:
        """Apply zero-noise extrapolation."""
        try:
            scaled_states = []
            
            # Generate scaled noise versions
            for scale in self.scale_factors:
                scaled_state = self._apply_scaled_noise(state, scale)
                scaled_states.append(scaled_state)
            
            # Richardson extrapolation
            coefficients = self._richardson_coefficients(
                len(self.scale_factors)
            )
            
            extrapolated = np.zeros_like(state)
            for coeff, scaled_state in zip(coefficients, scaled_states):
                extrapolated += coeff * scaled_state
            
            # Apply holographic constraints
            extrapolated = self._apply_holographic_constraints(extrapolated)
            
            return extrapolated
            
        except Exception as e:
            logger.error(f"Zero-noise extrapolation failed: {str(e)}")
            raise
    
    def _probabilistic_error_cancellation(
        self,
        state: np.ndarray
    ) -> np.ndarray:
        """Apply probabilistic error cancellation."""
        try:
            # Generate quasi-probability distribution
            quasi_probs = self._generate_quasi_probabilities(state)
            
            # Sample from quasi-probability distribution
            n_samples = 1000
            cancelled_state = np.zeros_like(state)
            
            for _ in range(n_samples):
                # Sample operation
                op_idx = np.random.choice(
                    len(quasi_probs),
                    p=np.abs(quasi_probs)/np.sum(np.abs(quasi_probs))
                )
                
                # Apply operation with sign
                sign = np.sign(quasi_probs[op_idx])
                cancelled_state += sign * self._apply_operation(
                    state,
                    op_idx
                )
            
            cancelled_state /= n_samples
            
            # Apply holographic constraints
            cancelled_state = self._apply_holographic_constraints(
                cancelled_state
            )
            
            return cancelled_state
            
        except Exception as e:
            logger.error(f"Error cancellation failed: {str(e)}")
            raise
    
    def _quasi_probability_decomposition(
        self,
        state: np.ndarray
    ) -> np.ndarray:
        """Apply quasi-probability decomposition."""
        try:
            # Generate basis operations
            basis_ops = self._generate_basis_operations()
            
            # Optimize decomposition
            def objective(params):
                reconstructed = np.zeros_like(state)
                for p, op in zip(params, basis_ops):
                    reconstructed += p * op @ state
                return np.linalg.norm(reconstructed - state)
            
            result = minimize(
                objective,
                np.zeros(len(basis_ops)),
                method='SLSQP',
                constraints={'type': 'eq', 'fun': lambda x: np.sum(x) - 1}
            )
            
            # Apply optimal decomposition
            decomposed = np.zeros_like(state)
            for p, op in zip(result.x, basis_ops):
                decomposed += p * op @ state
            
            # Apply holographic constraints
            decomposed = self._apply_holographic_constraints(decomposed)
            
            return decomposed
            
        except Exception as e:
            logger.error(f"Quasi-probability decomposition failed: {str(e)}")
            raise
    
    def _apply_holographic_constraints(
        self,
        state: np.ndarray
    ) -> np.ndarray:
        """Apply holographic constraints to quantum state."""
        try:
            # Ensure normalization
            state /= np.sqrt(np.sum(np.abs(state)**2))
            
            # Apply entropy bound
            density = np.abs(state)**2
            entropy = -np.sum(density * np.log2(density + 1e-10))
            
            if entropy > self.n_qubits:
                # Project onto maximum entropy state
                state *= np.sqrt(2**-self.n_qubits / density)
            
            return state
            
        except Exception as e:
            logger.error(f"Constraint application failed: {str(e)}")
            raise 